+++ pydocx/export/mixins/__init__.py
@@ -1,0 +1,7 @@
+from pydocx.export.mixins.faked_superscript_and_subscript import (
+    FakedSuperscriptAndSubscriptExportMixin,
+)
+
+__all__ = [
+    'FakedSuperscriptAndSubscriptExportMixin',
+]